+++ client/views/room/ContextualBar/OTR/OTR.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import VerticalBar from '../basic/VerticalBar';
+import VerticalBar from '../../../../components/basic/VerticalBar';
 import { OTR } from './OTR';
 
 export default {
