+++ client/views/room/ContextualBar/OTR/index.js
@@ -1,0 +1,3 @@
+import OTR from './OTR';
+
+export default OTR;
